fix(range): demo拆解与规范
