+++ src/packages/range/demo.tsx
@@ -1,28 +1,21 @@
-import React, { useState } from 'react'
-import { Range } from './range'
-import { Cell } from '../cell/cell'
+import React from 'react'
 import { useTranslate } from '../../sites/assets/locale'
-import Toast from '../toast'
-import ConfigProvider from '../configprovider'
-
-interface T {
-  title: string
-  title1: string
-  title2: string
-  title3: string
-  title4: string
-  title5: string
-  title6: string
-  title7: string
-  title8: string
-  title9: string
-  title10: string
-  title11: string
-  controlled: string
-}
+import Demo1 from './demos/h5/demo1'
+import Demo2 from './demos/h5/demo2'
+import Demo3 from './demos/h5/demo3'
+import Demo4 from './demos/h5/demo4'
+import Demo5 from './demos/h5/demo5'
+import Demo6 from './demos/h5/demo6'
+import Demo7 from './demos/h5/demo7'
+import Demo8 from './demos/h5/demo8'
+import Demo9 from './demos/h5/demo9'
+import Demo10 from './demos/h5/demo10'
+import Demo11 from './demos/h5/demo11'
+import Demo12 from './demos/h5/demo12'
+import Demo13 from './demos/h5/demo13'
 
 const RangeDemo = () => {
-  const [translated] = useTranslate<T>({
+  const [translated] = useTranslate({
     'zh-CN': {
       title: '基础用法',
       title1: '双滑块',
@@ -54,206 +47,36 @@ const RangeDemo = () => {
       controlled: 'Controlled',
     },
   })
-  const cellStyle = {
-    padding: '40px 18px',
-  }
-  const verticalStyle = {
-    height: '180px',
-    padding: '10px',
-  }
-  const [value1, setValue1] = useState(40)
-  const [value2, setValue2] = useState(60)
-  const [marks] = useState({
-    0: 'Start',
-    20: 20,
-    40: 40,
-    60: 60,
-    80: 80,
-    100: 'End',
-  })
 
   return (
     <>
       <div className="demo">
         <h2>{translated.title}</h2>
-        <Cell style={cellStyle}>
-          <Range defaultValue={40} onEnd={(val) => Toast.show(`${val}`)} />
-        </Cell>
-        <Cell style={cellStyle}>
-          <Range
-            defaultValue={40}
-            marks={[
-              { value: 0, label: 'start' },
-              { value: 100, label: 'end' },
-            ]}
-            onEnd={(val) => Toast.show(`${val}`)}
-          />
-        </Cell>
+        <Demo1 />
         <h2>{translated.controlled}</h2>
-        <Cell style={cellStyle}>
-          <Range value={value1} onChange={(val: any) => setValue1(val)} />
-        </Cell>
+        <Demo2 />
         <h2>{translated.title11}</h2>
-        <Cell style={cellStyle}>
-          <Range
-            defaultValue={40}
-            minDescription="0%"
-            maxDescription="100%"
-            currentDescription={(value) => `${value}%`}
-            onEnd={(val) => Toast.show(`${val}`)}
-          />
-        </Cell>
+        <Demo3 />
         <h2>{translated.title1}</h2>
-        <Cell style={cellStyle}>
-          <Range
-            defaultValue={[20, 80]}
-            range
-            onEnd={(val) => Toast.show(`${val}`)}
-          />
-        </Cell>
+        <Demo4 />
         <h2>{translated.title2}</h2>
-        <Cell style={cellStyle}>
-          <Range
-            defaultValue={0}
-            max={10}
-            min={-10}
-            onEnd={(val) => Toast.show(`${val}`)}
-          />
-        </Cell>
+        <Demo5 />
         <h2>{translated.title3}</h2>
-        <Cell style={cellStyle}>
-          <Range
-            defaultValue={30}
-            step={5}
-            onEnd={(val) => Toast.show(`${val}`)}
-          />
-        </Cell>
+        <Demo6 />
         <h2>{translated.title4}</h2>
-        <Cell style={cellStyle}>
-          <Range
-            defaultValue={30}
-            maxDescription={null}
-            minDescription={null}
-            onEnd={(val) => Toast.show(`${val}`)}
-          />
-        </Cell>
+        <Demo7 />
         <h2>{translated.title5}</h2>
-        <Cell style={cellStyle}>
-          <Range
-            defaultValue={20}
-            currentDescription={null}
-            onEnd={(val) => Toast.show(`${val}`)}
-          />
-        </Cell>
+        <Demo8 />
         <h2>{translated.title6}</h2>
-        <Cell style={cellStyle}>
-          <Range defaultValue={50} disabled />
-        </Cell>
+        <Demo9 />
         <h2>{translated.title7}</h2>
-        <Cell
-          style={{
-            ...cellStyle,
-            display: 'block',
-          }}
-        >
-          <ConfigProvider
-            theme={{
-              nutuiRangeButtonBorder: '1px solid rgba(52,96,250,1)',
-              nutuiRangeActiveColor:
-                'linear-gradient(315deg, rgba(73,143,242,1) 0%,rgba(73,101,242,1) 100%)',
-              nutuiRangeInactiveColor: 'rgba(163,184,255,1)',
-              nutuiRangeMargin: '20px',
-              nutuiRangeHeight: '6px',
-            }}
-          >
-            <Range
-              className="test-range"
-              defaultValue={40}
-              style={{ color: 'red' }}
-              marks={{
-                10: 10,
-                20: 20,
-              }}
-            />
-          </ConfigProvider>
-        </Cell>
+        <Demo10 />
         <h2>{translated.title8}</h2>
-        <Cell style={cellStyle}>
-          <Range
-            value={value2}
-            button={
-              <div
-                style={{
-                  width: '26px',
-                  color: 'white',
-                  fontSize: '10px',
-                  lineHeight: '18px',
-                  textAlign: 'center',
-                  backgroundColor: 'red',
-                  borderRadius: '10px',
-                }}
-              >
-                {value2}
-              </div>
-            }
-            onChange={(val: any) => setValue2(val)}
-            onEnd={(val) => Toast.show(`${val}`)}
-          />
-        </Cell>
+        <Demo11 />
         <h2>{translated.title9}</h2>
-        <Cell style={verticalStyle}>
-          <div style={{ width: '150px', height: '100%' }}>
-            <Range
-              defaultValue={20}
-              vertical
-              onEnd={(val) => Toast.show(`${val}`)}
-            />
-          </div>
-          <div style={{ width: '150px', height: '100%' }}>
-            <Range
-              defaultValue={[20, 80]}
-              vertical
-              range
-              onEnd={(val) => Toast.show(`${val}`)}
-            />
-          </div>
-        </Cell>
+        <Demo12 />
         <h2>{translated.title10}</h2>
-        <Cell style={cellStyle}>
-          <Range
-            defaultValue={60}
-            maxDescription={null}
-            minDescription={null}
-            marks={marks}
-            onEnd={(val) => Toast.show(`${val}`)}
-          />
-        </Cell>
-        <Cell style={cellStyle}>
-          <Range
-            defaultValue={[20, 80]}
-            marks={marks}
-            range
-            onEnd={(val) => Toast.show(`${val}`)}
-          />
-        </Cell>
-
-        <Cell style={verticalStyle}>
-          <Range
-            defaultValue={60}
-            vertical
-            maxDescription={null}
-            minDescription={null}
-            marks={marks}
-            onEnd={(val) => Toast.show(`${val}`)}
-          />
-          <Range
-            defaultValue={[20, 80]}
-            vertical
-            marks={marks}
-            range
-            onEnd={(val) => Toast.show(`${val}`)}
-          />
-        </Cell>
+        <Demo13 />
       </div>
     </>
   )
