+++ src/packages/range/demos/taro/demo1.tsx
@@ -1,0 +1,40 @@
+import React, { useState } from 'react'
+import { Range, Cell, Toast } from '@nutui/nutui-react-taro'
+
+const Demo1 = () => {
+  const cellStyle = {
+    padding: '40px 18px',
+  }
+  const [show, setShow] = useState(false)
+  const [msg, setMsg] = useState('')
+  const showToast = (msg: string) => {
+    setMsg(msg)
+    setShow(true)
+  }
+  return (
+    <>
+      <Cell style={cellStyle}>
+        <Range defaultValue={40} onEnd={(val) => showToast(`${val}`)} />
+      </Cell>
+      <Cell style={cellStyle}>
+        <Range
+          defaultValue={40}
+          marks={[
+            { value: 0, label: 'start' },
+            { value: 100, label: 'end' },
+          ]}
+          onEnd={(val) => showToast(`${val}`)}
+        />
+      </Cell>
+      <Toast
+        type="text"
+        visible={show}
+        content={msg}
+        onClose={() => {
+          setShow(false)
+        }}
+      />
+    </>
+  )
+}
+export default Demo1
